Move cache to it's own file
